test: split deal ticket account tests
